Fix NVIZ fly mode on 64bit nixes. Fixes https://trac.osgeo.org/grass/ticket/46. Thanks to Ken Kwasnicki.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@42592 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ visualization/nviz/src/togl_flythrough.c
@@ -316,7 +316,7 @@ double this_time(void)
     struct timeval tv;
 
     gettimeofday(&tv, NULL);
-    return ((float)tv.tv_sec + ((float)tv.tv_usec / 1000000.0));
+    return ((double)tv.tv_sec + ((double)tv.tv_usec / 1000000.0));
 #endif
 }
